+++ packages/localizations/src/he-IL.ts
@@ -606,6 +606,12 @@ export const heIL: LocalizationResource = {
       subtitle: 'להמשיך אל {{applicationName}}',
       title: 'בדוק את הדוא"ל שלך',
     },
+    emailCodeMfa: {
+      formTitle: 'בדוק את הדוא"ל שלך',
+      resendButton: 'לא קיבלת קוד? שלח שוב',
+      subtitle: 'להמשיך אל {{applicationName}}',
+      title: 'בדוק את הדוא"ל שלך',
+    },
     emailLink: {
       clientMismatch: {
         subtitle: 'להמשך, פתח את קוד האימות מהמכשיר והדפדפן ממנו אתה מתכוון לבצע כניסה',
@@ -641,6 +647,12 @@ export const heIL: LocalizationResource = {
         titleNewTab: 'נכנס בכרטיסייה אחרת',
       },
     },
+    emailLinkMfa: {
+      formSubtitle: 'השתמש בקישור האימות שנשלח לדוא״ל שלך',
+      resendButton: 'לא קיבלת את הקישור? שלח שוב',
+      subtitle: 'כדי להמשיך ל-{{applicationName}}',
+      title: 'בדוק את הדוא״ל שלך',
+    },
     enterpriseConnections: {
       subtitle: undefined,
       title: undefined,
@@ -658,6 +670,7 @@ export const heIL: LocalizationResource = {
       label__alternativeMethods: 'או, התחבר באמצעות שיטה אחרת.',
       title: 'שכחת סיסמה?',
     },
+    newDeviceVerificationNotice: 'אתה נכנס מכשיר חדש. אנו מבקשים אימות כדי לשמור על החשבון שלך מאובטח.',
     noAvailableMethods: {
       message: 'לא ניתן להמשיך בהתחברות. אין גורם אימות זמין.',
       subtitle: 'אירעה שגיאה',
@@ -1036,7 +1049,7 @@ export const heIL: LocalizationResource = {
       title: 'הוסף חשבון מחובר',
     },
     deletePage: {
-      actionDescription: 'הקלד "מחק חשבון" למטה כדי להמשיך.',
+      actionDescription: 'הקלד "מחיקת חשבון" למטה כדי להמשיך.',
       confirm: 'מחק חשבון',
       messageLine1:
         'האם אתה בטוח שברצונך למחוק את החשבון שלך? חלק מהנתונים המשויכים עשויים להישמר. כדי לבקש מחיקת נתונים מלאה, אנא צור קשר עם התמיכה.',
@@ -1278,16 +1291,16 @@ export const heIL: LocalizationResource = {
   },
   waitlist: {
     start: {
-      actionLink: undefined,
-      actionText: undefined,
-      formButton: undefined,
-      subtitle: undefined,
-      title: undefined,
+      actionLink: 'התחבר',
+      actionText: 'כבר יש לך גישה?',
+      formButton: 'הצטרף לרשימת ההמתנה',
+      subtitle: 'הזן את כתובת הדוא"ל שלך ואנו נודיע לך כשהמקום שלך יהיה מוכן',
+      title: 'הצטרף לרשימת ההמתנה',
     },
     success: {
-      message: undefined,
-      subtitle: undefined,
-      title: undefined,
+      message: 'תועבר בקרוב...',
+      subtitle: 'אנו ניצור איתך קשר כשהמקום שלך יהיה מוכן',
+      title: 'תודה שהצטרפת לרשימת ההמתנה!',
     },
   },
 } as const;