+++ packages/localizations/src/ro-RO.ts
@@ -622,6 +622,12 @@ export const roRO: LocalizationResource = {
       subtitle: 'pentru a continua la {{applicationName}}',
       title: 'Verifică-ți emailul',
     },
+    emailCodeMfa: {
+      formTitle: 'Verifică-ți emailul',
+      resendButton: 'Nu ai primit un cod? Retrimite',
+      subtitle: 'pentru a continua la {{applicationName}}',
+      title: 'Verifică-ți emailul',
+    },
     emailLink: {
       clientMismatch: {
         subtitle:
@@ -658,6 +664,12 @@ export const roRO: LocalizationResource = {
         titleNewTab: 'Autentificat în altă filă',
       },
     },
+    emailLinkMfa: {
+      formSubtitle: 'Folosiți linkul de verificare trimis la adresa dvs. de e-mail',
+      resendButton: 'Nu ați primit linkul? Retrimiteți',
+      subtitle: 'pentru a continua către {{applicationName}}',
+      title: 'Verificați-vă e-mailul',
+    },
     enterpriseConnections: {
       subtitle: undefined,
       title: undefined,
@@ -675,6 +687,8 @@ export const roRO: LocalizationResource = {
       label__alternativeMethods: 'Sau autentifică-te cu altă metodă',
       title: 'Ai uitat parola?',
     },
+    newDeviceVerificationNotice:
+      'Te conectezi de pe un dispozitiv nou. Solicităm verificarea pentru a menține contul tău sigur.',
     noAvailableMethods: {
       message: 'Nu se poate continua autentificarea. Nu există niciun factor de autentificare disponibil.',
       subtitle: 'A apărut o eroare',